# api/openapi.yaml
openapi: 3.0.0
info:
  title: CTF Management API
  description: 'API for managing CTF (Capture The Flag) games, teams, users, and services.'
  contact: {}
  version: 1.1.0
servers:
  - url: 'https://ctf01d.ru'
    description: Production server
  - url: 'http://localhost:4102'
    description: Local server
...
paths:
...
  /api/v1/teams:
    get:
      tags:
        - Teams
      summary: List all teams
      operationId: listTeams
      parameters: []
      responses:
        '200':
          description: A list of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
                x-content-type: application/json
        '400':
          description: Failed request
          content:
            application/json:
              schema: {}
        '500':
          description: Failed response
          content:
            application/json:
              schema: {}
    post:
      tags:
        - Teams
      summary: Create a new team
      operationId: createTeam
      security:
        - sessionAuth: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '200':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Failed request
          content:
            application/json:
              schema: {}
        '500':
          description: Failed response
          content:
            application/json:
              schema: {}
  '/api/v1/teams/{teamId}':
    get:
      tags:
        - Teams
      summary: Get a team by ID
      operationId: getTeamById
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Detailed information of a team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
    put:
      tags:
        - Teams
      summary: Update a team
      operationId: updateTeam
      security:
        - sessionAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '200':
          description: Team updated successfully
    delete:
      tags:
        - Teams
      summary: Delete a team
      operationId: deleteTeam
      security:
        - sessionAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Team deleted successfully
  '/api/v1/teams/{teamId}/members':
    get:
      tags:
        - Teams
      summary: Get all members of a team
      operationId: teamMembers
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '404':
          description: Team not found
  '/api/v1/teams/{teamId}/requests/{userId}':
    post:
      tags:
        - Teams
      summary: Create request to join team
      operationId: connectUserTeamRequest
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      responses:
        '200':
          description: Request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "User request to join team created successfully"
    put:
      tags:
        - Teams
      summary: Approve user's request to join the team
      operationId: approveUserTeamRequest
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Request approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "User approved and added to team successfully"
    delete:
      tags:
        - Teams
      summary: Reject user's request to join the team
      operationId: rejectUserTeamRequest
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Request rejected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "User request rejected successfully"
  '/api/v1/teams/{teamId}/members/{userId}':
    delete:
      tags:
        - Teams
      summary: Remove member from team
      operationId: leaveUserFromTeam
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "User removed from team successfully"
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier for the user
        display_name:
          type: string
          description: The name of the user
        user_name:
          type: string
          description: The login of the user
        role:
          type: string
          enum:
            - admin
            - player
            - guest
          example: player
          description: 'The role of the user (admin, player or guest)'
        avatar_url:
          type: string
          description: URL to the user's avatar
        status:
          type: string
          description: 'Status of the user (active, disabled)'
...
    TeamRequest:
      required:
        - name
        - university_id
      type: object
      properties:
        name:
          type: string
          description: Name of the team
        description:
          type: string
          description: A brief description of the team
        university_id:
          type: string
          format: uuid
          description: University or institution the team is associated with
        social_links:
          type: string
          description: JSON string containing social media links of the team
        avatar_url:
          type: string
          description: URL to the team's avatar
    TeamResponse:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the team
        name:
          type: string
          description: Name of the team
        description:
          type: string
          description: A brief description of the team
        university:
          type: string
          description: University or institution the team is associated with
        social_links:
          type: string
          description: JSON string containing social media links of the team
        avatar_url:
          type: string
          description: URL to the team's avatar
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
security: []